# [TraceQL] Improve performance of select() queries