@mdisibio 2024-12-12
Bumped to -count=10 and benchmarks show less variance, so good indicator it is noise.  Nothing changed with the fetch or engine for anything except select.

<details>
<summary>Benchmarks</summary>

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/tempo/tempodb/encoding/vparquet3
cpu: Apple M4 Pro
                                                 │  before.txt  │              after.txt              │
                                                 │    sec/op    │   sec/op     vs base                │
BackendBlockTraceQL/spanAttValMatch                 67.69m ± 2%   66.99m ± 2%        ~ (p=0.481 n=10)
BackendBlockTraceQL/spanAttValNoMatch               1.564m ± 2%   1.546m ± 1%   -1.11% (p=0.000 n=10)
BackendBlockTraceQL/spanAttIntrinsicMatch           37.34m ± 1%   37.31m ± 1%        ~ (p=0.353 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch         1.216m ± 0%   1.205m ± 0%   -0.91% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttValMatch             292.5m ± 1%   293.0m ± 2%        ~ (p=0.315 n=10)
BackendBlockTraceQL/resourceAttValNoMatch           1.216m ± 1%   1.211m ± 1%        ~ (p=0.315 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch       10.03m ± 1%   10.05m ± 1%        ~ (p=0.739 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch#01    1.196m ± 0%   1.188m ± 0%   -0.73% (p=0.000 n=10)
BackendBlockTraceQL/traceOrMatch                    149.3m ± 1%   149.3m ± 1%        ~ (p=1.000 n=10)
BackendBlockTraceQL/traceOrNoMatch                  143.2m ± 1%   143.3m ± 1%        ~ (p=0.579 n=10)
BackendBlockTraceQL/mixedValNoMatch                 92.26m ± 0%   92.09m ± 0%   -0.18% (p=0.015 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd           1.348m ± 1%   1.346m ± 1%        ~ (p=0.631 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr            91.35m ± 1%   91.57m ± 1%        ~ (p=0.436 n=10)
BackendBlockTraceQL/count                           239.3m ± 1%   238.2m ± 0%        ~ (p=0.105 n=10)
BackendBlockTraceQL/struct                          271.8m ± 3%   268.4m ± 1%        ~ (p=0.105 n=10)
BackendBlockTraceQL/||                              106.9m ± 1%   106.3m ± 1%        ~ (p=0.123 n=10)
BackendBlockTraceQL/mixed                           30.30m ± 1%   29.95m ± 2%   -1.18% (p=0.015 n=10)
BackendBlockTraceQL/complex                         1.760m ± 0%   1.709m ± 0%   -2.92% (p=0.000 n=10)
BackendBlockTraceQL/select                         200.78m ± 2%   15.62m ± 1%  -92.22% (p=0.000 n=10)
geomean                                             25.03m        21.76m       -13.05%

                                                 │  before.txt  │                after.txt                │
                                                 │     B/s      │      B/s       vs base                  │
BackendBlockTraceQL/spanAttValMatch                212.0Mi ± 2%    214.3Mi ± 2%          ~ (p=0.481 n=10)
BackendBlockTraceQL/spanAttValNoMatch              1.068Gi ± 2%    1.080Gi ± 1%     +1.12% (p=0.000 n=10)
BackendBlockTraceQL/spanAttIntrinsicMatch          401.4Mi ± 1%    401.7Mi ± 1%          ~ (p=0.353 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch        957.8Mi ± 0%    966.6Mi ± 0%     +0.92% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttValMatch            48.38Mi ± 1%    48.29Mi ± 2%          ~ (p=0.305 n=10)
BackendBlockTraceQL/resourceAttValNoMatch          141.9Mi ± 1%    142.4Mi ± 1%          ~ (p=0.315 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch      1.227Gi ± 1%    1.223Gi ± 1%          ~ (p=0.756 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch#01   173.0Mi ± 0%    174.3Mi ± 0%     +0.74% (p=0.000 n=10)
BackendBlockTraceQL/traceOrMatch                   87.08Mi ± 1%    87.11Mi ± 1%          ~ (p=0.986 n=10)
BackendBlockTraceQL/traceOrNoMatch                 6.056Mi ± 1%    6.051Mi ± 1%          ~ (p=0.616 n=10)
BackendBlockTraceQL/mixedValNoMatch                21.48Mi ± 0%    21.52Mi ± 0%     +0.18% (p=0.021 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd          137.2Mi ± 1%    137.4Mi ± 1%          ~ (p=0.617 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr           167.0Mi ± 1%    166.6Mi ± 1%          ~ (p=0.436 n=10)
BackendBlockTraceQL/count                          59.08Mi ± 1%    59.34Mi ± 0%          ~ (p=0.101 n=10)
BackendBlockTraceQL/struct                         58.04Mi ± 3%    58.76Mi ± 1%          ~ (p=0.105 n=10)
BackendBlockTraceQL/||                             133.0Mi ± 1%    133.8Mi ± 1%          ~ (p=0.123 n=10)
BackendBlockTraceQL/mixed                          461.5Mi ± 1%    467.0Mi ± 2%     +1.20% (p=0.015 n=10)
BackendBlockTraceQL/complex                        872.2Mi ± 2%    893.1Mi ± 1%     +2.39% (p=0.000 n=10)
BackendBlockTraceQL/select                         70.90Mi ± 2%   911.57Mi ± 1%  +1185.79% (p=0.000 n=10)
geomean                                            155.7Mi         179.0Mi         +14.97%

                                                 │ before.txt  │              after.txt               │
                                                 │  MB_io/op   │  MB_io/op    vs base                 │
BackendBlockTraceQL/spanAttValMatch                 15.05 ± 0%    15.05 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/spanAttValNoMatch               1.793 ± 0%    1.793 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/spanAttIntrinsicMatch           15.72 ± 0%    15.72 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/spanAttIntrinsicNoMatch         1.221 ± 0%    1.221 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttValMatch             14.84 ± 0%    14.84 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttValNoMatch          180.9m ± 0%   180.9m ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttIntrinsicMatch       13.21 ± 0%    13.21 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttIntrinsicMatch#01   217.1m ± 0%   217.1m ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/traceOrMatch                    13.63 ± 0%    13.63 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/traceOrNoMatch                 909.1m ± 0%   909.1m ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixedValNoMatch                 2.078 ± 0%    2.078 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixedValMixedMatchAnd          193.9m ± 0%   193.9m ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixedValMixedMatchOr            16.00 ± 0%    16.00 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/count                           14.82 ± 0%    14.82 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/struct                          16.54 ± 0%    16.54 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/||                              14.91 ± 0%    14.91 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixed                           14.66 ± 0%    14.66 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/complex                         1.613 ± 1%    1.599 ± 1%       ~ (p=0.159 n=10)
BackendBlockTraceQL/select                          14.93 ± 0%    14.93 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                             4.086         4.084       -0.04%
¹ all samples are equal

                                                 │   before.txt   │              after.txt               │
                                                 │      B/op      │     B/op      vs base                │
BackendBlockTraceQL/spanAttValMatch                  29.64Mi ± 1%   29.64Mi ± 0%        ~ (p=0.912 n=10)
BackendBlockTraceQL/spanAttValNoMatch                1.539Mi ± 0%   1.539Mi ± 0%        ~ (p=0.445 n=10)
BackendBlockTraceQL/spanAttIntrinsicMatch            14.80Mi ± 0%   14.79Mi ± 0%        ~ (p=0.353 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch          1.096Mi ± 0%   1.096Mi ± 0%        ~ (p=0.191 n=10)
BackendBlockTraceQL/resourceAttValMatch              316.9Mi ± 0%   316.9Mi ± 0%        ~ (p=0.853 n=10)
BackendBlockTraceQL/resourceAttValNoMatch            1.071Mi ± 0%   1.071Mi ± 0%        ~ (p=0.892 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch        1.321Mi ± 0%   1.322Mi ± 0%        ~ (p=0.381 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch#01     1.073Mi ± 0%   1.073Mi ± 0%   -0.00% (p=0.019 n=10)
BackendBlockTraceQL/traceOrMatch                     12.72Mi ± 1%   12.71Mi ± 0%        ~ (p=0.093 n=10)
BackendBlockTraceQL/traceOrNoMatch                   11.17Mi ± 0%   11.17Mi ± 0%        ~ (p=0.986 n=10)
BackendBlockTraceQL/mixedValNoMatch                  1.579Mi ± 0%   1.579Mi ± 0%        ~ (p=1.000 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd            1.074Mi ± 0%   1.074Mi ± 0%        ~ (p=0.880 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr             2.093Mi ± 1%   2.115Mi ± 2%        ~ (p=0.165 n=10)
BackendBlockTraceQL/count                            231.4Mi ± 0%   231.6Mi ± 0%        ~ (p=0.393 n=10)
BackendBlockTraceQL/struct                           12.11Mi ± 4%   11.98Mi ± 3%        ~ (p=0.631 n=10)
BackendBlockTraceQL/||                               16.41Mi ± 0%   16.42Mi ± 0%        ~ (p=0.225 n=10)
BackendBlockTraceQL/mixed                            4.208Mi ± 0%   4.209Mi ± 0%        ~ (p=0.912 n=10)
BackendBlockTraceQL/complex                          1.139Mi ± 0%   1.140Mi ± 0%        ~ (p=0.796 n=10)
BackendBlockTraceQL/select                         113.185Mi ± 2%   1.675Mi ± 0%  -98.52% (p=0.000 n=10)
geomean                                              6.664Mi        5.339Mi       -19.88%

                                                 │  before.txt  │               after.txt               │
                                                 │  allocs/op   │  allocs/op   vs base                  │
BackendBlockTraceQL/spanAttValMatch                 313.6k ± 0%   313.6k ± 0%        ~ (p=0.562 n=10)
BackendBlockTraceQL/spanAttValNoMatch               17.30k ± 0%   17.30k ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/spanAttIntrinsicMatch           118.4k ± 0%   118.4k ± 0%        ~ (p=0.415 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch         17.23k ± 0%   17.23k ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttValMatch             2.293M ± 0%   2.293M ± 0%        ~ (p=0.986 n=10)
BackendBlockTraceQL/resourceAttValNoMatch           17.26k ± 0%   17.26k ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttIntrinsicMatch       18.55k ± 0%   18.55k ± 0%        ~ (p=1.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch#01    17.26k ± 0%   17.26k ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/traceOrMatch                    120.2k ± 0%   120.2k ± 0%        ~ (p=0.953 n=10)
BackendBlockTraceQL/traceOrNoMatch                  111.7k ± 0%   111.7k ± 0%        ~ (p=0.870 n=10)
BackendBlockTraceQL/mixedValNoMatch                 18.03k ± 0%   18.03k ± 0%        ~ (p=1.000 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd           17.30k ± 0%   17.30k ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixedValMixedMatchOr            24.17k ± 0%   24.17k ± 0%        ~ (p=0.903 n=10)
BackendBlockTraceQL/count                           1.476M ± 0%   1.476M ± 0%   +0.00% (p=0.003 n=10)
BackendBlockTraceQL/struct                          52.71k ± 1%   53.38k ± 1%   +1.29% (p=0.015 n=10)
BackendBlockTraceQL/||                              141.6k ± 0%   141.7k ± 0%        ~ (p=0.259 n=10)
BackendBlockTraceQL/mixed                           47.56k ± 0%   47.56k ± 0%        ~ (p=1.000 n=10)
BackendBlockTraceQL/complex                         17.76k ± 0%   17.76k ± 0%        ~ (p=0.169 n=10)
BackendBlockTraceQL/select                         942.62k ± 2%   22.88k ± 0%  -97.57% (p=0.000 n=10)
geomean                                             70.33k        57.87k       -17.72%
¹ all samples are equal
```
</details>